% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bodenuebersichts.R
\docType{data}
\name{Bodenuebersichts}
\alias{Bodenuebersichts}
\title{Bodenuebersichts}
\format{
A data frame with three columns and 19183 rows
\describe{
  \item{x}{coordinates in x dimensions with 1000 m resoultion}
  \item{y}{coordinates in y dimensions with 1000 m resoultion}
  \item{TKLE_NR}{values describing the soil map types in SpatRaster}
}
}
\source{
Bodenuebersichtskarte 1:200.000 (BueK200) \url{https://www.bgr.bund.de/DE/Themen/Boden/Projekte/Informationsgrundlagen-laufend/BUEK200/BUEK200.html}
}
\usage{
Bodenuebersichts
}
\description{
Soil maps types for KlimaKonfrom Model domain
as a dataframe with similar resolution as Monica model.
}
\keyword{datasets}
