% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clima_diagramm_change.R
\name{clima_diagramm_change}
\alias{clima_diagramm_change}
\title{Climate Diagram (Relative)}
\usage{
clima_diagramm_change(data, temp_precip_mean, location, language, run_id)
}
\arguments{
\item{data}{A data frame containing the data to be plotted. The data should contain
min, max, and mean temperature and precipitation.}

\item{temp_precip_mean}{A vector of the long term average of temperature and precipitation.}

\item{location}{A character referring to the location of the data, e.g., "Greiz".}

\item{language}{A character variable either "EN" or "DE", to specify the languages of plots.}

\item{run_id}{A character variable either pointing out the run id, please refer
to the simulation setup file.}
}
\value{
ggplot2 plot
}
\description{
Plot the relative climate diagram for grid points from the climate data used to run the Monica models.
The function receives a data frame, preferable prepared by the R code
'/production/Prod_klima_diagramm_4corners.R'. The input data are representing grids
in the KlimaKonform model region or at least Regions. All meta data required to plot this are either
the production script or the function here.
}
\seealso{
clima_diagramm_change
}
\author{
Ahmed Homoudi
}
