% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clima_dMI.R
\name{clima_dMI}
\alias{clima_dMI}
\title{Calculate dMI (De Martonne Aridity Index) from ReKIS dataset}
\usage{
clima_dMI(
  csv.file,
  resolution = c("Jahr (Jan-Dez)", "Vegetationsperiode I (April-Juni)",
    "Vegetationsperiode II (Juli-Sept)", "Okt-Dez", "Jan-Mrz"),
  plot.type = c("ts", "bp"),
  axis.scales,
  output_path
)
}
\arguments{
\item{csv.file}{description}

\item{resolution}{description}

\item{plot.type}{description}

\item{axis.scales}{description}

\item{output_path}{description}
}
\value{
A PNG
}
\description{
This function calculates the dMI (De Martonne Aridity Index)) for climate
station data.
}
