% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clima_dMI_multiple.R
\name{clima_dMI_multiple}
\alias{clima_dMI_multiple}
\title{Calculate dMI (De Martonne Aridity Index) from ReKIS multiple stations}
\usage{
clima_dMI_multiple(
  csv.files,
  resolution = c("Jahr (Jan-Dez)", "Vegetationsperiode I (April-Juni)",
    "Vegetationsperiode II (Juli-Sept)"),
  plot.type = c("ts", "bp"),
  axis.scales,
  output_path
)
}
\arguments{
\item{csv.files}{description}

\item{resolution}{description}

\item{plot.type}{description}

\item{axis.scales}{description}

\item{output_path}{description}
}
\value{
A PNG
}
\description{
This function calculates the dMI (De Martonne Aridity Index)) for climate
station data.
}
