% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clima_Heliothermal_Index.R
\name{clima_Heliothermal_Index}
\alias{clima_Heliothermal_Index}
\title{Plotting Heliothermal Index}
\usage{
clima_Heliothermal_Index(x.df, lat)
}
\arguments{
\item{x.df}{A data frame containing Datum, tmax (daily max temp),
tmean (daily mean temp)}

\item{lat}{A single value of latitude}
}
\value{
ggplot2 plot
}
\description{
Plotting Heliothermal Index
}
\author{
Ahmed Homoudi
}
