% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_12maps_relative_change.R
\name{ens_12maps_relative_change}
\alias{ens_12maps_relative_change}
\title{Plotting 12 maps (3 historical + 9 relative change) from 3 RCP scenarios}
\usage{
ens_12maps_relative_change(
  netCDF.files,
  variable,
  region,
  landcover,
  stat_var,
  language,
  run_id,
  output_path,
  output_csv
)
}
\arguments{
\item{netCDF.files}{The netCDF files that contain similar simulation
variables. They should be provided including the full path to the files}

\item{variable}{inside these netCDf files, e.g., AET}

\item{region}{A string referring to the region to plot; "total",
"Vogtlandkreis", "Burgenlandkreis", "Greiz", and "Altenburger Land".}

\item{landcover}{A numeric variable indicating which land cover should be
considered. To consider all land cover give 1000.}

\item{stat_var}{A string referring to the statistical variable to plot in
case the file is ensemble file, for example, "mean","sd","median","max"
and "min"}

\item{language}{A character variable either "EN" or "DE", to specify the
languages of plots}

\item{run_id}{A character variable either "2ter", "3ter", or "4ter"
specifying the which run is used to produce the ensemble.}

\item{output_path}{A string pointing to the output directory,
it should end with "/"}

\item{output_csv}{A string pointing to the output directory for the csv
file that is produced by pre-processing netCDF files}
}
\value{
PNG
}
\description{
A function that receives three netCDF files, and variable name,
statistical ensemble member and plot 12 maps, considering 3 RCPs X four periods
it in the desired the out put directory in which the output is placed.
the first three maps represents the mean of the historical period, and the rest
represents the relative change to the historical period
}
\author{
Ahmed Homoudi
}
