% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_all_mems_heatmap.R
\name{ens_all_mems_heatmap}
\alias{ens_all_mems_heatmap}
\title{Plotting heatmaps of the three RCPs scenarios plus five ensembles}
\usage{
ens_all_mems_heatmap(
  netCDF.files,
  variable,
  region,
  landcover,
  language,
  run_id,
  output_path,
  output_csv
)
}
\arguments{
\item{netCDF.files}{The netCDF files that contain similar simulation
variables. They should be provided including the full path to the files}

\item{variable}{inside these netCDF files, e.g., AET}

\item{region}{A string referring to the region to plot; "total",
"Vogtlandkreis", "Burgenlandkreis", "Greiz", and "Altenburger Land".}

\item{landcover}{A numeric variable indicating which land cover should be
considered. To consider all land cover give 1000.}

\item{language}{A character variable either "EN" or "DE", to specify the
languages of plots}

\item{run_id}{A character variable either "2ter", "3ter", or "4ter"
specifying the which run is used to produce the ensemble.}

\item{output_path}{A string pointing to the output directory,
it should end with "/"}

\item{output_csv}{A string pointing to the output directory for the csv
file that is produced by pre-processing netCDF files}
}
\value{
PNG
}
\description{
A function that receives three netCDF files, and variable name,
and plot three plots heat maps
}
\author{
Ahmed Homoudi
}
