% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_public_output.R
\name{ens_public_output}
\alias{ens_public_output}
\title{Plotting one informative plot containing, relative change maps,
box plots and time series.}
\usage{
ens_public_output(
  netCDF.file,
  variable,
  region,
  landcover,
  stat_var,
  language,
  run_id,
  output_path,
  output_csv
)
}
\arguments{
\item{netCDF.file}{The netCDF file that contain similar simulation
variables. It should be provided including the full path to the files}

\item{variable}{inside these netCDf file, e.g., AET}

\item{region}{A string referring to the region to plot; "total",
"Vogtlandkreis", "Burgenlandkreis", "Greiz", and "Altenburger Land".}

\item{landcover}{A numeric variable indicating which land cover should be
considered. To consider all land cover give 1000.}

\item{stat_var}{A string referring to the statistical variable to plot in
case the file is ensemble file, for example, "mean","sd","median","max"
and "min"}

\item{language}{A character variable either "EN" or "DE", to specify the
languages of plots}

\item{run_id}{A character variable either "2ter", "3ter", or "4ter"
specifying the which run is used to produce the ensemble.}

\item{output_path}{A string pointing to the output directory,
it should end with "/"}

\item{output_csv}{A string pointing to the output directory for the csv
file that is produced by pre-processing netCDF files}
}
\value{
PNG
}
\description{
A function that receives one netCDF files, and variable name,
statistical ensemble member and plot 4 maps, box plots, & time series considering one RCPs.
it in the desired the out put directory in which the output is placed.
the first three maps represents the mean of the historical period, and the rest
represents the relative change to the historical period
}
\author{
Ahmed Homoudi
}
