% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_boxplots.R
\name{exp_boxplots}
\alias{exp_boxplots}
\title{Plotting boxplots from experiment files}
\usage{
exp_boxplots(
  netCDF.files,
  variable,
  region,
  landcover,
  y.axis.limits = NA,
  language,
  run_id,
  output_path,
  output_csv = NULL
)
}
\arguments{
\item{netCDF.files}{The netCDF files that contain similar simulation
variables. They should be provided including the full path to the files}

\item{variable}{inside these netCDF files, e.g., AET}

\item{region}{A string referring to the region to plot; "total",
"Vogtlandkreis", "Burgenlandkreis", "Greiz", and "Altenburger Land".}

\item{landcover}{A numeric variable indicating which land cover should be
considered. To consider all land cover give 1000.}

\item{y.axis.limits}{A vector of length 2L to produce all plots with similar
y axis limits.}

\item{language}{A character variable either "EN" or "DE", to specify the
languages of plots}

\item{run_id}{A character variable either "2ter", "3ter", "4ter", or "6ter"
#' specifying the which run is used to produce the ensemble.}

\item{output_path}{A string pointing to the output directory,
it should end with "/"}

\item{output_csv}{A string pointing to the output directory for the csv
file that is produced by pre-processing netCDF files}
}
\value{
PNG
}
\description{
A function that receives three netCDF files, and variable name,
statistical ensemble member and plot boxplots of 4 different period and place
it  in the desired the out put directory in which the output is placed.
}
\author{
Ahmed Homoudi
}
