% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KlimaKonformC3-package.R
\docType{package}
\name{KlimaKonformC3-package}
\alias{KlimaKonformC3}
\alias{KlimaKonformC3-package}
\title{KlimaKonformC3: Analysis of Agricultural Simulations, KlimaKonform (C3)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Analysis the Agricultural Simulations performed within of the Projekt KlimaKonform (TB C3 Landwirtschaft).
}
\author{
\strong{Maintainer}: Ahmed Homoudi \email{ahmed.homoudi@tu-dresden.de} (\href{https://orcid.org/0000-0001-6906-5986}{ORCID})

}
\keyword{internal}
