% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landkreise.R
\docType{data}
\name{Landkreise}
\alias{Landkreise}
\title{Political districts shape file}
\format{
A tibble with 3 rows and 2 variables:
\describe{
  \item{Name}{Name of the political district}
  \item{geometry}{geometry of the political districte}
}
}
\source{
Bundesamt für Kartographie und Geodäsie - 2021 \url{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-250-000-ebenen-stand-01-01-vg250-ebenen-01-01.html}
}
\usage{
Landkreise
}
\description{
Data of shapefile of the political districts will be used for plotting
Included the political districts: Vogtlandkreis, Burgenlandkreis, and Greiz
}
\keyword{datasets}
