% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ensemble_from_sims.R
\name{make_ensemble_from_sims}
\alias{make_ensemble_from_sims}
\title{Making an ensemble from individual simulations (i.e. individual simulation as netCDf files)}
\usage{
make_ensemble_from_sims(nc.files, variable, output_path)
}
\arguments{
\item{nc.files}{The netCDF files that contain similar simulation variables.
They should be provided including the full path to the files}

\item{variable}{inside these netCDf files, e.g., AET}

\item{output_path}{A string pointing to the output directory, it should end with "/"}
}
\value{
netCDf file
}
\description{
A function that receives multiple netCDF files, and variable name,
the desired the output directory in which the output is placed.
}
\author{
Ahmed Homoudi
}
