% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round_custom}
\alias{round_custom}
\title{A function to receive a number and round up or down according to the next number indicated by roundTo}
\source{
\url{https://stackoverflow.com/a/32508105/13818750}
}
\usage{
round_custom(x, roundTo, dir = 1)
}
\arguments{
\item{x}{A number to round}

\item{roundTo}{number indicating to which number to round, e.g., 5 , or 10}

\item{dir}{A binary indicating rounding  down zero or up one}
}
\value{
rounded number
}
\description{
A function to receive a number and round up or down according to the next number indicated by roundTo
}
\examples{
x <- c(25, 50, 75, 125, 150, 175, 220)
roundTo <- 5
dir <- 1
round_custom(
  x,
  roundTo,
  dir
)
}
\author{
theforestecologist in StackOverflow
}
