% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_linear_trend.R
\name{sim_linear_trend}
\alias{sim_linear_trend}
\title{Plotting linear trend from one simulation froom KlimaKonform TB C3.}
\usage{
sim_linear_trend(
  netCDF.file,
  region,
  landcover,
  y.axis.limits,
  language,
  run_id,
  output_path,
  output_csv
)
}
\arguments{
\item{netCDF.file}{The netCDF file provided including the full path to the file}

\item{region}{A string referring to the region to plot; "total","Vogtlandkreis",
"Burgenlandkreis", "Greiz", and "Altenburger Land".}

\item{landcover}{A numeric variable indicating which land cover should be considered.
To consider all land cover give 1000.}

\item{y.axis.limits}{A vector of length 2L to produce all plots with similar y axis limits.
If not provided, each variable has being supplied with limits in the package.}

\item{language}{A character variable either "EN" or "DE", to specify the languages of plots}

\item{run_id}{A character variable either "2ter", "3ter", or "4ter" specificng the which run is used
to produce the ensemble.}

\item{output_path}{A string pointing to the output directory, it should end with "/"}

\item{output_csv}{A string pointing to the output directory for the csv file that is produced by
pre-processing netCDF file}
}
\value{
PNG
}
\description{
A function that receives one netCDF files, variable name,
and, custom y axis limits and plot linear trend plot.
The desired the out put directory in which the output is placed.
}
\author{
Ahmed Homoudi
}
