% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_output_de.R
\docType{data}
\name{standard_output_de}
\alias{standard_output_de}
\title{Plotting Standards in German}
\format{
A tibble with 10 rows and 12 variables:
\describe{
  \item{pattern}{dbl Year price was recorded}
  \item{longname}{long name of the variable}
  \item{units}{units of the variable}
  \item{variable}{variable}
  \item{standardname}{standardname}
  \item{frequency}{frequency}
  \item{comment}{comment}
  \item{lowe_limit}{lowe_limit}
  \item{upper_limit}{upper_limit}
  \item{interval}{interval}
}
}
\source{
The KlimaKonform Convection
}
\usage{
standard_output_de
}
\description{
Data of standards will be used for plotting in German
}
\keyword{datasets}
