% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_output_en.R
\docType{data}
\name{standard_output_en}
\alias{standard_output_en}
\title{Plotting Standards in English}
\format{
A tibble with 10 rows and 12 variables:
\describe{
\item{pattern}{dbl Year price was recorded}
\item{longname}{long name of the variable}
\item{units}{units of the variable}
\item{variable}{variable}
\item{standardname}{standardname}
\item{frequency}{frequency}
\item{comment}{comment}
}
}
\source{
The KlimaKonform Convection
}
\usage{
standard_output_en
}
\description{
Data of standards will be used for plotting in English
}
\keyword{datasets}
