% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clima_diagramm_change.R
\name{clima_diagramm_change}
\alias{clima_diagramm_change}
\title{Climate Diagram (Relative)}
\usage{
clima_diagramm_change(data, run_id, output_path)
}
\arguments{
\item{data}{A data frame containing the daily data to be plotted. The data should contain
min, max, and mean temperature and precipitation. For only three locations, e.g., MO, MM, and MU.}

\item{run_id}{A character variable either pointing out the run id, please refer
to the simulation setup file.}

\item{output_path}{A string pointing to the output directory for output plot}
}
\value{
PNG
}
\description{
Plot the relative change as climate diagram for grid points from the climate
data used to run the Monica model. The function receives a data frame, preferable prepared by the
R code '/production/Prod_klima_diagramm_4corners.R'. The input data are representing grids
the KlimaKonform model region. All meta data required to plot this are either
the production script or the function here.
}
\seealso{
clima_diagramm_abs
}
\author{
Ahmed Homoudi
}
